Add Javascript solution to Problem 1
I used grunt to create a javascript project skeleton and wrote
my code in the provided framework. I'm not sure exactly how much
of the formatting I used was required and how much was just
cargo-culting, but I'll sort that out in time. In the meantime,
the tests pass and the code makes sense.